Move KeybaseNotifications class into its own file, create a KeybaseNotification elasticsearch model, and actually save a notification to the database on user_registered
